html/canvas/element/transformations/2d.transformation.rotateZ.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.transformation.rotateZ</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.transformation.rotateZ</h1>
<p class="desc">rotateZ() results in the correct transformation matrix</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("rotateZ() results in the correct transformation matrix");
_addTest(function(canvas, ctx) {

// angles are in radians, test something that is not a multiple of pi
const angle = 2;
const domMatrix = new DOMMatrix();
ctx.rotateZ(angle);
domMatrix.rotateAxisAngleSelf(0, 0, 1, angle / Math.PI * 180);
let canvasMatrix = ctx.getTransform();
_assert(domMatrix.toLocaleString() == canvasMatrix.toLocaleString(), "domMatrix.toLocaleString() == canvasMatrix.toLocaleString()");
ctx.rotateZ(angle);
domMatrix.rotateAxisAngleSelf(0, 0, 1, angle / Math.PI * 180);
canvasMatrix = ctx.getTransform();
_assert(domMatrix.toLocaleString() == canvasMatrix.toLocaleString(), "domMatrix.toLocaleString() == canvasMatrix.toLocaleString()");


});
</script>
